Move Sort.Order to separate class, add AutoGson, rename getItem to getId
