+++ molgenis-api/src/main/java/org/molgenis/api/model/Sort.java
@@ -7,7 +7,7 @@ import com.google.auto.value.AutoValue;
 import java.util.List;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nullable;
-import org.molgenis.api.model.Sort.Order.Direction;
+import org.molgenis.api.model.Order.Direction;
 
 @AutoValue
 public abstract class Sort {
@@ -29,44 +29,6 @@ public abstract class Sort {
 
   public static Builder builder() {
     return new AutoValue_Sort.Builder();
-  }
-
-  @AutoValue
-  public abstract static class Order {
-    public enum Direction {
-      ASC,
-      DESC
-    }
-
-    public abstract String getItem();
-
-    @Nullable
-    @CheckForNull
-    public abstract Direction getDirection();
-
-    public static Order create(String newItem) {
-      return create(newItem, null);
-    }
-
-    public static Order create(String newItem, @Nullable @CheckForNull Direction newDirection) {
-      return builder().setItem(newItem).setDirection(newDirection).build();
-    }
-
-    public static Builder builder() {
-      return new AutoValue_Sort_Order.Builder();
-    }
-
-    @SuppressWarnings(
-        "squid:S1610") // Abstract classes without fields should be converted to interfaces
-    @AutoValue.Builder
-    public abstract static class Builder {
-
-      public abstract Builder setItem(String newItem);
-
-      public abstract Builder setDirection(@Nullable @CheckForNull Direction newDirection);
-
-      public abstract Order build();
-    }
   }
 
   @SuppressWarnings(
